01.24

+++ backend/src/main/java/com/restaurant/backend/repository/RoleRepository.java
@@ -4,5 +4,5 @@ package com.restaurant.backend.repository;
 import com.restaurant.backend.model.Role;
 import org.springframework.data.repository.CrudRepository;
 
-public interface RoleRepository extends CrudRepository<Role, Long> {
+public interface RoleRepository extends CrudRepository<Role, Integer> {
 }
